requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.5.2
certifi==2025.1.31
channels==4.2.0
charset-normalizer==3.4.1
colorama==0.4.6
defusedxml==0.7.1
Django==5.1.6
django-allauth==65.5.0
django-filter==24.3
django-modelcluster==6.4
django-otp==1.5.4
django-otp-twilio==1.0.5
django-permissionedforms==0.1
django-smtp-ssl==1.0
django-stubs-ext==5.1.2
django-taggit==6.1.0
django-tasks==0.6.1
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et_xmlfile==2.0.0
filetype==1.2.0
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.1
httplib2==0.22.0
idna==3.10
laces==0.1.2
lxml==5.3.2
oauthlib==3.2.2
openpyxl==3.1.5
pillow==11.1.0
pillow_heif==0.21.0
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.22.0
pyparsing==3.2.1
python-docx==1.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
soupsieve==2.6
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wagtail==6.4.1
Willow==1.9.0
YubiOTP==1.0.0
